feat: マスタデータ管理を独立アプリ admin-tool へ分離
frontend の Clerk 認証から切り離し、ゲームバランス調整用のローカル専用ツールとして
admin-tool/（ポート3100）を新設。既存の /api/admin/* API と ADMIN_API_KEY 認証は
そのまま利用し、frontend 側の旧 admin 実装(pages/components/hooks/types/middleware)は削除した。
