+++ admin-tool/.env.local.example
@@ -1,0 +1,6 @@
+# Backend API URL
+# ローカル開発: http://127.0.0.1:8000
+NEXT_PUBLIC_API_URL=http://127.0.0.1:8000
+
+# Admin API Key（backend/.env の ADMIN_API_KEY と一致させること）
+NEXT_PUBLIC_ADMIN_API_KEY=your_admin_api_key_here
