+++ admin-tool/next.config.ts
@@ -1,0 +1,11 @@
+import path from "node:path";
+import type { NextConfig } from "next";
+
+const nextConfig: NextConfig = {
+  reactCompiler: true,
+  turbopack: {
+    root: path.join(__dirname),
+  },
+};
+
+export default nextConfig;
