+++ admin-tool/src/components/admin/MobileSuitEditForm.tsx
@@ -5,7 +5,7 @@ import { useEffect } from "react";
 import { useForm, useFieldArray, Controller } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { z } from "zod";
-import { MasterMobileSuit } from "@/types/battle";
+import { MasterMobileSuit } from "@/types/admin";
 
 // ============================================================
 // Zod バリデーションスキーマ
