+++ admin-tool/src/components/ui/index.ts
@@ -1,0 +1,3 @@
+export { default as SciFiPanel } from "./SciFiPanel";
+export { default as SciFiButton } from "./SciFiButton";
+export { default as SciFiHeading } from "./SciFiHeading";
