+++ frontend/src/types/battle.ts
@@ -11,4 +11,3 @@ export * from "./skill";
 export * from "./shop";
 export * from "./leaderboard";
 export * from "./social";
-export * from "./admin";
